chore: renamend "Genral" to "Global"
