+++ client/src/routes/dashboard/pages/rankings/GlobalRankingTable.tsx
