+++ client/src/routes/dashboard/pages/rankings/components/GlobalRankings.tsx
